% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFerroptosis.R
\name{getFerroptosis}
\alias{getFerroptosis}
\title{Get the Ferroptosis phenotype and FRRS}
\usage{
getFerroptosis(data = data)
}
\arguments{
\item{data}{The matrix of gene expression data, rows are samples and columns are genes.}
}
\value{
return a dataframe, the first column is the ID of each sample, the second column is the
Ferroptosis phenotype of each sample, and the third column is the FRRS of each sample.
}
\description{
Use gene expression data to get the corresponding Ferroptosis phenotype and Ferroptosis related risk score of each patient
}
\details{
Make sure that the data you enter is a matrix, and rows are samples and columns are genes.
}
