% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTIME.R
\name{getTIME}
\alias{getTIME}
\title{Get the TIME phenotype and TIMEscore}
\usage{
getTIME(data = data)
}
\arguments{
\item{data}{The matrix of gene expression data, rows are samples and columns are genes.}
}
\value{
return a dataframe, the first column is the ID of each sample, the second column is the
TIME phenotype of each sample, and the third column is the TIME index of each sample.
}
\description{
Use gene expression data to get the corresponding TIME phenotype and TIMEscore of each patient
}
\details{
Make sure that the data you enter is a matrix, and rows are samples and columns are genes.
}
